+++ devel/ags/ags.cpp
@@ -25,17 +25,21 @@
 
 using namespace ags;
 
-void aggregated_storage_driver__::_Free( void )
+void aggregated_storage_driver__::_Free(bso::sBool EvenIfPersistent)
 {
-	if ( _Descriptor != AGS_UNDEFINED_DESCRIPTOR )
-		_AStorage->Free( _Descriptor );
+	if ( ( _Descriptor != AGS_UNDEFINED_DESCRIPTOR ) && ( EvenIfPersistent || _AStorage->IsVolatile() ) )
+		_AStorage->Free( _Descriptor ); // Freeing must not be done on a 'reset(true)' with a persistent storage.
+}
+
+sdr::eType aggregated_storage_driver__::SDRType(void) const
+{
+  return _AStorage->Type();
 }
 
 void aggregated_storage_driver__::SDRAllocate( sdr::size__ Size )
 {
 	_Descriptor = _AStorage->Reallocate( _Descriptor, Size );
 }
-
 
 sdr::size__ aggregated_storage_driver__::SDRSize( void ) const
 {
@@ -45,12 +49,15 @@ sdr::size__ aggregated_storage_driver__::SDRSize( void ) const
 		return 0;
 }
 
-void aggregated_storage_driver__::SDRRecall(
+sdr::sSize aggregated_storage_driver__::SDRFetch(
 	sdr::row_t__ Position,
 	sdr::size__ Amount,
-	sdr::byte__ *Buffer )
+	sdr::byte__ *Buffer,
+	qRPN)
 {
 	_AStorage->Read( _Descriptor, Position, Amount, Buffer );
+
+	return Amount;
 }
 // lit  partir de 'Position' et place dans 'Tampon' 'Nombre' octets;
 void aggregated_storage_driver__::SDRStore(