+++ devel/uys/uys.h
@@ -127,23 +127,32 @@ namespace uys {
 #endif
 			return _Driver;
 		}
+		sdr::eType Type(void) const
+		{
+		  _Test();
+
+		  return _Driver->Type();
+		}
 		void Allocate( sdr::size__ Size )
 		{
 			_Test();
 
 			_Driver->Allocate( Size );
 		}
-		void Recall(
+		sdr::sSize Fetch(
 			sdr::row_t__ Position,
 			sdr::size__ Amount,
-			sdr::byte__ *Buffer ) const
+			sdr::byte__ *Buffer,
+			qRPN) const
 		{
 			_Test();
 
 			if ( _CVMBuffer != NULL )
 				memcpy( Buffer, _CVMBuffer + Position, Amount );
 			else
-				_Driver->Recall( Position, Amount, Buffer );
+				return _Driver->Fetch(Position, Amount, Buffer, qRP);
+
+      return Amount;
 		}
 		void Store(
 			const sdr::byte__ *Buffer,
@@ -199,13 +208,14 @@ namespace uys {
 			if ( Amount > ( S_.Size - Position ) )
 				qRFwk();
 		}
-		void _Recall(
+		sdr::sSize _Fetch(
 			sdr::row_t__ Position,
 			sdr::size__ Amount,
-			sdr::byte__ *Buffer ) const
+			sdr::byte__ *Buffer,
+			qRPN) const
 		{
 			_Test( Position, Amount );
-			_Driver.Recall( Position, Amount, Buffer );
+			return _Driver.Fetch(Position, Amount, Buffer, qRP);
 		}
 		void _Store(
 			const sdr::byte__ *Buffer,
@@ -285,19 +295,24 @@ namespace uys {
 			sdr::row_t__ Position,
 			sdr::size__ Size );
 		//f Allocates 'Capacity' bytes.
+		sdr::eType Type(void) const
+		{
+		  return _Driver.Type();
+		}
 		void Allocate( sdr::size__ Size )
 		{
 			_Allocate( Size );
 		}
-		/*f Recall 'Amount' bytes at 'Position' and put them in 'Buffer'.
+		/*f Fetch 'Amount' bytes at 'Position' and put them in 'Buffer'.
 		Ignore is only for 'UYS_DBG' mode and for the 'MMG' library.
 		When 'true', it didn't make the test about the size. */
-		void Recall(
+		sdr::sSize Fetch(
 			sdr::row_t__ Position,
 			sdr::size__ Amount,
-			sdr::byte__ *Buffer ) const
+			sdr::byte__ *Buffer,
+			qRPN) const
 		{
-			_Recall( Position, Amount, Buffer );
+			return _Fetch(Position, Amount, Buffer, qRP);
 		}
 		//f Store 'Amount' bytes from 'Buffer' at 'Offset'.
 		void Store(
@@ -308,18 +323,18 @@ namespace uys {
 			_Store( Buffer, Amount, Position );
 		}
 		//f Put byte at 'Position' in 'Datum'.
-		void Recall(
+		void Get(
 			sdr::row_t__ Position,
-			sdr::byte__ &Datum ) const
+			sdr::byte__ &Datum) const
 		{
-			Recall( Position, 1, &Datum );
+			Fetch(Position, 1, &Datum, qRPDefault);
 		}
 		//f Return byte at 'Position'.
 		sdr::byte__ Get( sdr::row_t__ Position ) const
 		{
 			sdr::byte__ D;
 
-			Recall( Position, D );
+			Get(Position, D);
 
 			return D;
 		}
@@ -350,7 +365,7 @@ namespace uys {
 				if ( TargetBuffer == NULL )
 					IndirectCopy_( Source, Offset, *this, Position, Amount );
 				else if ( Amount != 0 )
-					Source.Recall( Offset, Amount, TargetBuffer + Position );
+					Source.Fetch(Offset, Amount, TargetBuffer + Position, qRPDefault);
 			} else if ( TargetBuffer == NULL ) {
 				if ( Amount != 0  )
 					Store( SourceBuffer + Offset, Amount, Position );
@@ -433,15 +448,6 @@ namespace uys {
 
 		return m_Undefined;	// Pour viter un 'warning'.
 	}
-
-	qENUM( Behavior )
-	{
-		bVolatile,
-		bPersistent,
-		b_amount,
-		b_Undefined,
-		b_Default = bPersistent
-	};
 }
 
 namespace uys {
@@ -509,15 +515,15 @@ namespace uys {
 		eState Init(
 			const rHF &Filenames,
 			mode__ Mode,
-			eBehavior Behavior = b_Default,
+			sdr::eType Type = sdr::tPersistent,
 			flsq::rId Id = flsq::Undefined)
 		{
 			fsd::Init(Id, Filenames.Filename, Convert_( Mode ), flsq::cFirstUse);
 
-			switch ( Behavior ) {
-			case bVolatile:
+			switch ( Type ) {
+			case sdr::tVolatile:
 				break;
-			case bPersistent:
+			case sdr::tPersistent:
 				fsd::Persistent();
 				break;
 			default:
@@ -533,10 +539,10 @@ namespace uys {
 		eState Init(
 			const fnm::rName &Name,
 			mode__ Mode,
-			eBehavior Behavior = b_Default,
+			sdr::eType Type = sdr::tPersistent,
 			flsq::rId Id = flsq::Undefined)
 		{
-		  return Init(rHF(Name), Mode, Behavior, Id);
+		  return Init(rHF(Name), Mode, Type, Id);
 		}
 		eState State( void ) const
 		{
@@ -626,12 +632,15 @@ namespace uys {
 		void Init( void )
 		{}
 		//f Put in 'Buffer' 'Amount' bytes at 'Position'.
-		void Recall(
+		sdr::sSize Fetch(
 			sdr::row_t__ Position,
 			sdr::size__ Amount,
-			sdr::byte__ *Buffer ) const
+			sdr::byte__ *Buffer,
+			qRPType) const
 		{
 			memcpy( Buffer, m::Data_ + Position, Amount );
+
+			return Amount;
 		}
 		//f Write to 'Position' 'Amount' bytes from 'Buffer'.
 		void Store(
@@ -675,7 +684,7 @@ namespace uys {
 			sdr::row_t__ Position = 0,
 			sdr::row_t__ Offset = 0 )
 		{
-			Source.Recall( Position, Quantity, *m::Data_ + Offset );
+			Source.Fetch(Position, Quantity, *m::Data_ + Offset, qRPDefault);
 		}
 		//f Fill at 'Position' with 'Object' of size 'Size' 'Count' times.
 		void Store(