+++ stable/lstctn.h
@@ -218,13 +218,13 @@ namespace lstctn {
 		uys::eState Init(
 			const rHF &Filenames,
 			uys::mode__ Mode,
-			uys::eBehavior Behavior,
+			sdr::eType Type,
 			flsq::rId Id)
 		{
-			uys::eState State = Container_.Init(Filenames.Container, Mode, Behavior, Id);
+			uys::eState State = Container_.Init(Filenames.Container, Mode, Type, Id);
 
 			if ( !State.IsError() ) {
-				if ( List_.Init(Filenames.List, Mode, Behavior, Id) != State )
+				if ( List_.Init(Filenames.List, Mode, Type, Id) != State )
 					State = uys::sInconsistent;
 			}